Feat: criando metodos de finalização de apostas

+++ bolao/utils.py
@@ -1,10 +1,9 @@
 import os
 from secrets import token_hex
-from decimal import Decimal
 
 
 def get_taxa_banca():
-    return Decimal(os.getenv('TAXA_BANCA'))
+    return float(os.getenv('TAXA_BANCA'))
 
 
 def gerar_codigo():
